+++ gets.rb
@@ -1,0 +1,6 @@
+module Gets
+  def take_input_label(label)
+    print "#{label}: "
+    gets.chomp
+  end
+end
